Make it work
* Remove Schema namespace
* require Scenic::ActiveRecord::Statements
