+++ lib/scenic.rb
@@ -1,5 +1,6 @@
 require "scenic/version"
 require "scenic/engine"
+require "scenic/active_record/statements"
 
 module Scenic
 end
