+++ spec/scenic/active_record/statements_spec.rb
@@ -1,10 +1,10 @@
 require "spec_helper"
-require "scenic/active_record/schema"
+require "scenic/active_record/statements"
 
 class View < ActiveRecord::Base
 end
 
-describe "Scenic::ActiveRecord::Schema" do
+describe "Scenic::ActiveRecord::Statements" do
   describe "create_view" do
     it "creates a view from a file" do
       with_view_definition :views, 1, "SELECT text 'Hello World' AS hello" do
